[libromdata] CBMDOS::loadFieldData(): Add some size_t casts to silence some warnings.
CBMDOS.cpp: In member function ‘virtual int LibRomData::CBMDOS::loadFieldData()’:
CBMDOS.cpp:1135:28: warning: comparison of integer expressions of different signedness: ‘int’ and ‘std::vector<LibRomData::CBMDOSPrivate::track_offsets_t, rp::default_init_allocator<LibRomData::CBMDOSPrivate::track_offsets_t, std::allocator<LibRomData::CBMDOSPrivate::track_offsets_t> > >::size_type’ {aka ‘long unsigned int’} [-Wsign-compare]
 1135 |         if (d->dir_track-1 >= d->track_offsets.size()) {
      |             ~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/libromdata/Media/CBMDOS.cpp
@@ -1131,8 +1131,8 @@ int CBMDOS::loadFieldData(void)
 	auto vv_icons = new RomFields::ListDataIcons_t;	// for GEOS files only
 	bool has_icons = false;
 
-	assert(d->dir_track-1 < d->track_offsets.size());
-	if (d->dir_track-1 >= d->track_offsets.size()) {
+	assert((size_t)d->dir_track-1 < d->track_offsets.size());
+	if ((size_t)d->dir_track-1 >= d->track_offsets.size()) {
 		// Unable to read the directory track...
 		// TODO: Show an error?
 		return static_cast<int>(d->fields.count());
